test&fix: add logout and search.py
**ANOTHER IMPORTANT NOTE!!!**
```
Must use aysnc API instead of sync API,
otherwise
```
FAILED source/test/system-test/test_register.py::test_register - playwright._impl._api_types.Error: It looks like you are using Playwright Sync API inside the asyncio loop.
Please use the Async API instead.
```
will show up.
