+++ source/test/system-test/logout.py
@@ -1,7 +1,7 @@
-from playwright.async_api import Playwright, async_playwright, expect
+from playwright.sync_api import Playwright, sync_playwright, expect
 
 
-async def run(playwright: Playwright) -> None:
+def run(playwright: Playwright) -> None:
     browser = playwright.chromium.launch(headless=True)
     context = browser.new_context()
     page = context.new_page()
@@ -11,11 +11,12 @@ async def run(playwright: Playwright) -> None:
     page.get_by_label("帳號").press("Tab")
     page.locator("#password").fill("112522105")
     page.get_by_role("button", name="登入").click()
+    page.get_by_role("button", name="登出").click()
 
     # ---------------------
     context.close()
     browser.close()
 
-async def main():
-    async with async_playwright() as playwright:
-        await run(playwright)
+
+with sync_playwright() as playwright:
+    run(playwright)
