+++ source/test/system-test/register.py
@@ -1,6 +1,6 @@
-from playwright.sync_api import Playwright, sync_playwright, expect
+from playwright.async_api import Playwright, async_playwright, expect
 
-def run(playwright: Playwright) -> None:
+async def run(playwright: Playwright) -> None:
     browser = playwright.chromium.launch(headless=True)
     context = browser.new_context()
     page = context.new_page()
@@ -24,6 +24,7 @@ def run(playwright: Playwright) -> None:
     context.close()
     browser.close()
 
-with sync_playwright() as playwright:
-    browser = playwright.chromium.launch(headless=True)
-    run(playwright)
+async def main():
+    async with async_playwright() as playwright:
+        browser = await playwright.chromium.launch(headless=True)
+        await run(playwright)
