+++ source/test/system-test/search.py
@@ -1,0 +1,19 @@
+from playwright.sync_api import Playwright, sync_playwright, expect
+
+
+def run(playwright: Playwright) -> None:
+    browser = playwright.chromium.launch(headless=True)
+    context = browser.new_context()
+    page = context.new_page()
+    page.goto("http://localhost:8080/#/dashboard")
+    page.get_by_role("link", name="館藏查詢").click()
+    page.get_by_placeholder("Search").click()
+    page.get_by_placeholder("Search").fill("hi2")
+
+    # ---------------------
+    context.close()
+    browser.close()
+
+
+with sync_playwright() as playwright:
+    run(playwright)
